# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "resource-reservation-system-backend | v1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5007/"
        }
    ],
    "paths": {
        "/api/facility": {
            "post": {
                "tags": [
                    "Facility"
                ],
                "operationId": "e1c4d930f3e58f6a588bf0aa1e58bdf5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "capacity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/facility/get-facilities-info": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "operationId": "77b02d44ff94cdb38e84784030745b84",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/facility/get-facilities": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "operationId": "a683099aa02ca4eca602ead734300ef2",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/facility/get-facility-by-id/{id}": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "operationId": "835a947248f3dade47209aff4d2cafc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/create-reservation": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "37865bdaebd77f578db9f3469331aee5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "end": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "facilityId": {
                                        "type": "integer"
                                    },
                                    "firstName": {
                                        "type": "string"
                                    },
                                    "lastName": {
                                        "type": "string"
                                    },
                                    "phoneNumber": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/get-reservations": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "804fe0a05db4464f164365a7cf88c47f",
                "parameters": [
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/get-reservation/{id}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "f8babaed79ae193040060380b9b953c8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/get-reservation-by-facility/{facilityId}": {
            "get": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "25283ad8676a90f4891ac8afc38af541",
                "parameters": [
                    {
                        "name": "facilityId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/ongoing-reservation": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "22af182359c82ea027800bf49512d97f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/cancel-reservation": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "c9d22b77b17da77f20a21b69650b6a40",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/done-reservation": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "19a22360b347fc5c87091baacad6f460",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/reservation/move-reservation": {
            "post": {
                "tags": [
                    "Reservation"
                ],
                "operationId": "6ae7a52ec0262f1905de77c51c3c335f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "newStart": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "newEnd": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Facility",
            "description": "Facility"
        },
        {
            "name": "Reservation",
            "description": "Reservation"
        }
    ]
}